font url relative to stylesheet
Previously, font url was interpreted as relative to mapconfig url, which was intended to be compatible with javascript browser, which interprets font url as relative to the whole application.
This has become an issue due to the url inheriting protocol name, which might be unsuitable.

New behavior is to interpret font url as relative to the stylesheet it came from.

This change applies to bitmap urls as well.